add store selectors
