+++ src/components/movie-list/movie-list.connect.jsx
@@ -1,8 +1,10 @@
 import {connect} from "react-redux";
 import MoviesList from "@components/movie-list/movie-list";
+import {getMoviesByGenre} from "../../store/selectors";
 
 const mapStateToProps = (state) => ({
-  movies: state.filteredMovies
+  movies: getMoviesByGenre(state)
 });
 
 export default connect(mapStateToProps)(MoviesList);
+
