+++ src/components/movie-list/movie-list.jsx
@@ -76,4 +76,3 @@ MoviesList.propTypes = {
 };
 
 export default MoviesList;
-
