+++ src/constants/index.js
@@ -4,6 +4,8 @@ const MainMovie = {
   YEAR: 2014
 };
 
+const ALL_GENRES = `All genres`;
+
 const MOVIE_LAUNCH_DELAY = 1000;
 
-export {MainMovie, MOVIE_LAUNCH_DELAY};
+export {MainMovie, MOVIE_LAUNCH_DELAY, ALL_GENRES};
